Extracting core service wiring from RxPlatformSpace to separate Space
