+++ reflex-platform/src/hiconic/rx/platform/wire/space/RxPlatformSpace.java
@@ -1,34 +1,12 @@
 package hiconic.rx.platform.wire.space;
 
-import static com.braintribe.wire.api.util.Sets.set;
-
 import java.io.File;
-import java.util.Date;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
-import java.util.UUID;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.function.Supplier;
 
-import com.braintribe.codec.marshaller.common.BasicConfigurableMarshallerRegistry;
-import com.braintribe.codec.marshaller.json.JsonStreamMarshaller;
-import com.braintribe.codec.marshaller.yaml.YamlMarshaller;
-import com.braintribe.common.attribute.AttributeContext;
 import com.braintribe.gm.config.yaml.ModeledYamlConfiguration;
 import com.braintribe.gm.model.reason.Maybe;
 import com.braintribe.model.generic.GenericEntity;
 import com.braintribe.model.generic.reflection.EntityType;
-import com.braintribe.model.processing.service.common.ConfigurableDispatchingServiceProcessor;
-import com.braintribe.model.processing.service.common.context.UserSessionAspect;
-import com.braintribe.model.processing.service.common.eval.ConfigurableServiceRequestEvaluator;
-import com.braintribe.model.service.api.ServiceRequest;
-import com.braintribe.model.user.Role;
-import com.braintribe.model.user.User;
-import com.braintribe.model.usersession.UserSession;
-import com.braintribe.model.usersession.UserSessionType;
-import com.braintribe.utils.collection.impl.AttributeContexts;
 import com.braintribe.wire.api.annotation.Import;
 import com.braintribe.wire.api.annotation.Managed;
 import com.braintribe.wire.api.context.WireContext;
@@ -39,18 +17,13 @@ import hiconic.rx.module.api.wire.RxModuleContract;
 import hiconic.rx.module.api.wire.RxPlatformContract;
 import hiconic.rx.module.api.wire.RxProcessLaunchContract;
 import hiconic.rx.platform.loading.RxModuleLoader;
-import hiconic.rx.platform.models.RxCmdResolverManager;
-import hiconic.rx.platform.models.RxConfiguredModels;
 import hiconic.rx.platform.models.RxModelConfigurations;
-import hiconic.rx.platform.service.ContextualizingServiceRequestEvaluator;
 import hiconic.rx.platform.service.RxServiceDomain;
 import hiconic.rx.platform.service.RxServiceDomainConfigurations;
-import hiconic.rx.platform.service.RxServiceDomainDispatcher;
-import hiconic.rx.platform.service.RxServiceDomains;
 import hiconic.rx.platform.wire.contract.RxPlatformConfigContract;
 
 @Managed
-public class RxPlatformSpace implements RxPlatformContract, RxProcessLaunchContract {
+public class RxPlatformSpace extends CoreServicesSpace implements RxPlatformContract, RxProcessLaunchContract {
 
 	@Import
 	private RxPlatformConfigContract config;
@@ -116,90 +89,6 @@ public class RxPlatformSpace implements RxPlatformContract, RxProcessLaunchContr
 	}
 	
 	@Override
-	@Managed
-	public RxServiceDomains serviceDomains() {
-		RxServiceDomains bean = new RxServiceDomains();
-		bean.setContextEvaluator(evaluator());
-		bean.setFallbackProcessor(fallbackProcessor());
-		bean.setExecutorService(executorService());
-		bean.setModelConfigurations(modelConfigurations());
-		return bean;
-	}
-	
-	@Override
-	@Managed
-	public RxConfiguredModels configuredModels() {
-		RxConfiguredModels bean = new RxConfiguredModels();
-		bean.setCmdResolverManager(cmdResolverManager());
-		bean.setSystemAttributeContextSupplier(systemAttributeContextSupplier());
-		return bean;
-	}
-	
-	@Override
-	@Managed
-	public Supplier<AttributeContext> systemAttributeContextSupplier() {
-		return () -> AttributeContexts.derivePeek().set(UserSessionAspect.class, systemUserSession()).build();
-	}
-	
-	@Managed
-	public UserSession systemUserSession() {
-		UserSession bean = UserSession.T.create();
-
-		User user = systemUser();
-		
-		Set<String> effectiveRoles = new HashSet<>();
-		effectiveRoles.add("$all");
-		effectiveRoles.add("$user-" + user.getName());
-		for (Role userRole : user.getRoles())
-			effectiveRoles.add(userRole.getName());
-
-		Date now = new Date();
-
-		bean.setSessionId(UUID.randomUUID().toString());
-		bean.setType(UserSessionType.internal);
-		bean.setCreationInternetAddress("0:0:0:0:0:0:0:1");
-		bean.setCreationDate(now);
-		bean.setLastAccessedDate(now);
-		bean.setUser(user);
-		bean.setEffectiveRoles(effectiveRoles);
-
-		return bean;
-	}
-	
-	private static final Set<String> internalRoles = set("internal");
-	private static final String internalName = "internal";
-	
-	@Managed
-	public User systemUser() {
-
-		User bean = User.T.create();
-		bean.setId(internalName);
-		bean.setName(internalName);
-
-		for (String internalRoleName : internalRoles) {
-			Role internalRole = Role.T.create();
-			internalRole.setId(internalRoleName);
-			internalRole.setName(internalRoleName);
-			bean.getRoles().add(internalRole);
-		}
-
-		return bean;
-	}
-
-	@Managed
-	private RxCmdResolverManager cmdResolverManager() {
-		RxCmdResolverManager bean = new RxCmdResolverManager();
-		return bean;
-	}
-	
-	@Managed
-	public RxModelConfigurations modelConfigurations() {
-		RxModelConfigurations bean = new RxModelConfigurations();
-		bean.setConfiguredModels(configuredModels());
-		return bean;
-	}
-	
-	@Override
 	public <C extends GenericEntity> Maybe<C> readConfig(EntityType<C> configType) {
 		return modeledConfiguration().configReasoned(configType);
 	}
@@ -209,26 +98,6 @@ public class RxPlatformSpace implements RxPlatformContract, RxProcessLaunchContr
 		ModeledYamlConfiguration bean = new ModeledYamlConfiguration();
 		bean.setConfigFolder(new File(config.appDir(), "conf"));
 		return bean;
-	}
-	
-	@Managed
-	@Override
-	public BasicConfigurableMarshallerRegistry marshallers() {
-		BasicConfigurableMarshallerRegistry bean = new BasicConfigurableMarshallerRegistry();
-		bean.registerMarshaller("application/json", jsonMarshaller());
-		bean.registerMarshaller("text/yaml", yamlMarshaller());
-		bean.registerMarshaller("application/yaml", yamlMarshaller());
-		return bean;
-	}
-	
-	@Managed
-	private JsonStreamMarshaller jsonMarshaller() {
-		return new JsonStreamMarshaller();
-	}
-	
-	@Managed
-	private YamlMarshaller yamlMarshaller() {
-		return new YamlMarshaller();
 	}
 	
 	@Override
@@ -250,59 +119,5 @@ public class RxPlatformSpace implements RxPlatformContract, RxProcessLaunchContr
 	public String nodeId() {
 		return "main";
 	}
-	
-	@Override
-	@Managed
-	public ConfigurableServiceRequestEvaluator evaluator() {
-		ConfigurableServiceRequestEvaluator bean = new ConfigurableServiceRequestEvaluator();
-		bean.setExecutorService(executorService());
-		bean.setServiceProcessor(rootServiceProcessor());
-		return bean;
-	}
-	
-	@Override
-	public ContextualizingServiceRequestEvaluator evaluator(AttributeContext attributeContext) {
-		ContextualizingServiceRequestEvaluator bean = new ContextualizingServiceRequestEvaluator();
-		bean.setDelegate(evaluator());
-		bean.setAttributeContextProvider(() -> attributeContext);
-		return bean;
-	}
-	
-	@Override
-	@Managed
-	public ContextualizingServiceRequestEvaluator systemEvaluator() {
-		ContextualizingServiceRequestEvaluator bean = new ContextualizingServiceRequestEvaluator();
-		bean.setDelegate(evaluator());
-		bean.setAttributeContextProvider(systemAttributeContextSupplier());
-		return bean;
-	}
 
-	@Override
-	@Managed
-	public ExecutorService executorService() {
-		return Executors.newCachedThreadPool();
-	}
-
-	@Managed
-	public ConfigurableDispatchingServiceProcessor rootServiceProcessor() {
-		ConfigurableDispatchingServiceProcessor bean = new ConfigurableDispatchingServiceProcessor();
-
-		bean.register(ServiceRequest.T, serviceDomainDispatcher());
-		bean.registerInterceptor("domain-validation").register(serviceDomainDispatcher());
-		
-		return bean;
-	}
-	
-	@Managed
-	private RxServiceDomainDispatcher serviceDomainDispatcher() {
-		RxServiceDomainDispatcher bean = new RxServiceDomainDispatcher();
-		bean.setServiceDomains(serviceDomains());
-		return bean;
-	}
-	
-	@Managed
-	private ConfigurableDispatchingServiceProcessor fallbackProcessor() {
-		ConfigurableDispatchingServiceProcessor bean = new ConfigurableDispatchingServiceProcessor();
-		return bean;
-	}
 }
